Refactor project category configuration into a separate file
