+++ src/components/ProjectSummary.astro
@@ -1,6 +1,6 @@
 ---
 import { generateSlug } from "../data/projectFunctions";
-import { categoryConfig } from "../data/projects";
+import { categoryConfig } from "../data/projectsConfig";
 import type { Project } from "../data/projects";
 
 interface Props {
